DAOS-7951 test: Set DisableVMD in go unit tests
DisableVMD needs to be set by the caller in the Scan/Prepare/Format
requests. VMD is disabled by default in the server config, but will default
to false (meaning VMD will be enabled) if not manually set in the tests.

Currently unit tests will fail locally (and at some point in CI) if
VMD devices are present on the system. This will fix these test failures.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>